Handle duplicated constants causing warnings
Test the base implementation with shared behaviour and check
inheritance, only.
